requirements/base.txt
Django>=5.0
djangorestframework>=3.15
python-decouple>=3.8
Pillow>=12.0
djangorestframework-simplejwt>=5.0
django-redis>=5.0
redis>=5.0
django-ratelimit>=4.0
drf-spectacular>=0.27
httpx>=0.27
uvicorn>=0.30
channels==4.3.2
channels-redis==4.3.0
celery==5.6.3
flower==2.0.0
daphne==4.0.0